requirements.txt
bcrypt==4.1.2
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
dnspython==2.6.1
email-validator==2.1.1
flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
importlib-metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mysqlclient==2.2.4
packaging==24.0
pillow==10.2.0
pycparser==2.21
six==1.16.0
SQLAlchemy==2.0.28
typing-extensions==4.10.0
werkzeug==3.0.1
wtforms==3.1.2
zipp==3.18.1
